tests/wpt/web-platform-tests/html/cross-origin-embedder-policy/reporting-navigation.https.html

<!doctype html>
<html>
<body>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/common/get-host-info.sub.js"></script>
<script>
const {ORIGIN, REMOTE_ORIGIN} = get_host_info();
const COEP =
  '|header(cross-origin-embedder-policy,require-corp)' +
  '|header(cross-origin-embedder-policy-report-only,require-corp)';
const FRAME_URL = `${ORIGIN}/common/blank.html?pipe=`;
const REMOTE_FRAME_URL = `${REMOTE_ORIGIN}/common/blank.html?pipe=`;

function checkCorpReport(report, contextUrl, blockedUrl) {
  assert_equals(report.type, 'coep');
  assert_equals(report.url, contextUrl);
  assert_equals(report.body.type, 'corp');
  assert_equals(report.body['blocked-url'], blockedUrl);
}

function loadFrame(document, url) {
  return new Promise((resolve, reject) => {
    const frame = document.createElement('iframe');
    frame.src = url;
    frame.onload = () => resolve(frame);
    frame.onerror = reject;
    document.body.appendChild(frame);
  });
}

function observeReportsUpTo(global, upto) {
  const reports = [];
  return new Promise(resolve => {
    const observer = new global.ReportingObserver((rs) => {
      for (const r of rs) {
        reports.push(r.toJSON());
        if (reports.length == upto) {
          observer.disconnect();
          resolve(reports);
        }
      }
    });
    observer.observe();
  });
}

promise_test(async (t) => {
  const parent = await loadFrame(document, FRAME_URL + COEP);
  t.add_cleanup(() => parent.remove());
  loadFrame(parent.contentDocument, REMOTE_FRAME_URL + COEP);

  // One for COEP, one for COEP-RO.
  const reports = await observeReportsUpTo(parent.contentWindow, 2);

  assert_equals(reports.length, 2);
  checkCorpReport(reports[0], parent.src, REMOTE_FRAME_URL + COEP);
  checkCorpReport(reports[1], parent.src, REMOTE_FRAME_URL + COEP);
}, 'Navigation CORP');

promise_test(async (t) => {
  const grandParent = await loadFrame(document, FRAME_URL + COEP);
  t.add_cleanup(() => grandParent.remove());
  const parent = grandParent.contentDocument.createElement('iframe');
  grandParent.contentDocument.body.appendChild(parent);
  loadFrame(parent.contentDocument, REMOTE_FRAME_URL + COEP);

  // One for COEP, one for COEP-RO.
  const reports = await observeReportsUpTo(parent.contentWindow, 2);

  assert_equals(reports.length, 2);
  checkCorpReport(reports[0], 'about:blank', REMOTE_FRAME_URL + COEP);
  checkCorpReport(reports[1], 'about:blank', REMOTE_FRAME_URL + COEP);
}, 'Navigation CORP and about:blank');

</script>
</body></html>
